convert support file to TS
